# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API",
        "description": "API of sowgatly.app",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        },
        {
            "url": "https://sowgatly.app",
            "description": "Remote server"
        }
    ],
    "paths": {
...
        "/api/me/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "The caller's last 50 in-app notifications",
                "operationId": "f71fa8593bc3110e7334dedc31ba3856",
                "responses": {
                    "200": {
                        "description": "Notifications, newest first",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "order_created",
                                                            "order_status",
                                                            "product_available",
                                                            "chat_message",
                                                            "shop_application"
                                                        ]
                                                    },
                                                    "data": {
                                                        "description": "Keys depend on type, see docs/notifications.md",
                                                        "type": "object"
                                                    },
                                                    "title": {
                                                        "description": "Localized by Accept-Language",
                                                        "type": "string"
                                                    },
                                                    "body": {
                                                        "type": "string"
                                                    },
                                                    "read_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "unread": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/me/notifications/unread-count": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Number of unread in-app notifications",
                "operationId": "8cb5d181bf4ce1691fec49ba71382130",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "unread": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/me/notifications/read": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark every notification as read",
                "operationId": "6e6ba05cf6028e5fb32755deb5d7c2e7",
                "responses": {
                    "200": {
                        "description": "Done"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}